faster bit hack with int

+++ MedianFilterPlugin/medianfilterplugin.h
@@ -19,8 +19,8 @@ public:
     void edit(const cv::Mat &input, cv::Mat &output);
 
 private:
-    static void uint8_sort(uint8_t &a, uint8_t &b);
-    static void uint8_3x3_core_sort(std::array<uint8_t, 9> &a);
+    static void uint8_sort(int  &a, int  &b);
+    static void uint8_3x3_core_sort(std::array<int , 9> &a);
 };
 
 #endif // ROTATEPLUGIN_H
